Kobuk: Patch to enable ubus on server side.
Issue : Currently we are not able to do UCI SET from server code of
qcmap as it is running in radio

Fix :Enable ubus on server side to so, that we can
do UCI SET from server code.
But ubox and ubus lib's are in c. Require this patch to include
c lib from cpp code.

Change-Id: Iae2b889c3e661f3accc308fc4d998fad85c864ef
Signed-off-by: Dheeraj Kumar <quic_dhekum@quicinc.com>